requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
pandas>=1.3.0
scikit-learn>=0.24.0
python-docx>=0.8.11
PyPDF2>=2.0.0
python-dotenv>=0.19.0
nltk
spacy
matplotlib
pillow
